+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/_data_lake_file_client.py
@@ -586,6 +586,8 @@ class DataLakeFileClient(PathClient):
             and act according to the condition specified by the `match_condition` parameter.
         :keyword ~azure.core.MatchConditions match_condition:
             The match condition to use upon the etag.
+        :keyword bool decompress:
+            This is True by default. Set it to False to disable the file content auto-decompress.
         :keyword int max_concurrency:
             The number of parallel connections with which to download.
         :keyword int timeout: